+++ app/presentation/UnreadBadge.js
@@ -1,9 +1,29 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { View, Text } from 'react-native';
+import { View, Text, StyleSheet } from 'react-native';
 
-import styles from './styles';
-import { themes } from '../../constants/colors';
+import sharedStyles from '../views/Styles';
+import { themes } from '../constants/colors';
+
+const styles = StyleSheet.create({
+	unreadNumberContainer: {
+		minWidth: 21,
+		height: 21,
+		paddingVertical: 3,
+		paddingHorizontal: 5,
+		borderRadius: 10.5,
+		alignItems: 'center',
+		justifyContent: 'center',
+		marginLeft: 10
+	},
+	unreadText: {
+		overflow: 'hidden',
+		fontSize: 13,
+		...sharedStyles.textMedium,
+		letterSpacing: 0.56,
+		textAlign: 'center'
+	}
+});
 
 const UnreadBadge = React.memo(({
 	theme, unread, userMentions, groupMentions, style
@@ -36,7 +56,7 @@ const UnreadBadge = React.memo(({
 					styles.unreadText,
 					{ color }
 				]}
-			>{ unread }
+			>{unread}
 			</Text>
 		</View>
 	);
